decouple BVC+ShareActivity with BVC
# Conflicts:
#	ios/brave-ios/Sources/Brave/Frontend/Browser/BrowserViewController/BVC+ShareActivity.swift
