+++ ios/brave-ios/Sources/Brave/Frontend/Browser/BrowserViewController/BrowserViewController.swift
@@ -2710,23 +2710,27 @@ extension BrowserViewController: ToolbarUrlActionsDelegate {
         dismiss(animated: true) {
           self.presentActivityViewController(
             url,
-            sourceView: self.view,
-            sourceRect: self.view.convert(
-              self.topToolbar.shareButton.frame,
-              from: self.topToolbar.shareButton.superview
-            ),
-            arrowDirection: [.up]
+            source: .init(
+              view: self.view,
+              rect: self.view.convert(
+                self.topToolbar.shareButton.frame,
+                from: self.topToolbar.shareButton.superview
+              ),
+              arrowDirection: [.up]
+            )
           )
         }
       } else {
         presentActivityViewController(
           url,
-          sourceView: view,
-          sourceRect: view.convert(
-            topToolbar.shareButton.frame,
-            from: topToolbar.shareButton.superview
-          ),
-          arrowDirection: [.up]
+          source: .init(
+            view: self.view,
+            rect: self.view.convert(
+              self.topToolbar.shareButton.frame,
+              from: self.topToolbar.shareButton.superview
+            ),
+            arrowDirection: [.up]
+          )
         )
       }
     }
